feature(navigator) layout, element, component icon helpers fixed
