+++ editor/src/components/navigator/navigator-item/item-preview.tsx
@@ -2,7 +2,8 @@ import * as React from 'react'
 import { TemplatePath } from '../../../core/shared/project-file-types'
 import { IcnProps, Icn } from '../../../uuiui'
 import { betterReactMemo } from '../../../uuiui-deps'
-import { useNameAndIcon } from '../../inspector/common/name-and-icon-hook'
+import * as TP from '../../../core/shared/template-path'
+import { useComponentIcon } from '../layout-element-icons'
 
 interface ItemPreviewProps {
   path: TemplatePath
@@ -18,17 +19,21 @@ export const ItemPreview: React.FunctionComponent<ItemPreviewProps> = betterReac
     // 3 - if it's a component or not
     // 4 - if it's a placeholder or not
     // 5 if it's generated or not
-    const nameAndIcon = useNameAndIcon(props.path)
+    const iconProps = useComponentIcon(props.path)
 
-    return (
-      <div
-        className='w20 h20 flex justify-center items-center relative'
-        style={{
-          paddingLeft: 8,
-        }}
-      >
-        <Icn {...nameAndIcon.iconProps} color={props.color} />
-      </div>
-    )
+    if (iconProps == null || TP.isScenePath(props.path)) {
+      return null
+    } else {
+      return (
+        <div
+          className='w20 h20 flex justify-center items-center relative'
+          style={{
+            paddingLeft: 8,
+          }}
+        >
+          <Icn {...iconProps} color={props.color} />
+        </div>
+      )
+    }
   },
 )
